Crossflow effectiveness instead of CounterFlow
The dry effectiveness for single phase  fluid should be based on the
cross flow analysis. The equations are provided in the documentation of
ACHP page 52, but not implemented so far.
Now NTU_dry is updated for to handle cross flow instead of counter flow.
This is also seen in Christian Bach’s code.