Add new gas cooler model
The model was modified according to the work of Ortiz and Dang Li. In
supercritical region, the outlet temperature is iterated to satisfy
heat transfer between refrigerant and air side. At the same time, the
inner wall temperature is computed to update the heat transfer
coefficient (which is highly dependent on it). Dry Analysis-Single
phase was used instead of the condenser area analysis which improve the
accuracy of the gas cooler so much.